rsgi: make uvicorn dependencies optional
