+++ pyproject.toml
@@ -44,8 +44,6 @@ emmett = "emmett.cli:main"
 python = "^3.7"
 click = ">=6.0"
 granian = "0.1.0b2"
-h11 = ">= 0.12.0"
-h2 = ">= 3.2.0, < 4.1.0"
 pendulum = "~2.1.2"
 pyaes = "~1.6.1"
 pyDAL = "17.3"
@@ -53,14 +51,15 @@ python-rapidjson = "^1.0"
 pyyaml = "^5.4"
 renoir = "^1.5"
 severus = "^1.1"
-uvicorn = "~0.19.0"
-websockets = "^10.0"
-
-httptools = { version = "~0.5.0", markers = "sys_platform != 'win32'" }
-uvloop = { version = "~0.17.0", markers = "sys_platform != 'win32'" }
 
 orjson = { version = "~3.8", optional = true }
 emmett-crypto = { version = "^0.2.0", optional = true }
+
+uvicorn = { version = "^0.19.0", optional = true }
+h11 = { version = ">= 0.12.0", optional = true }
+h2 = { version = ">= 3.2.0, < 4.1.0", optional = true }
+websockets = { version = "^10.0", optional = true }
+httptools = { version = "~0.5.0", optional = true, markers = "sys_platform != 'win32'" }
 
 [tool.poetry.dev-dependencies]
 ipaddress = "^1.0"
@@ -72,6 +71,7 @@ psycopg2-binary = "^2.9.3"
 [tool.poetry.extras]
 crypto = ["emmett-crypto"]
 orjson = ["orjson"]
+uvicorn = ["uvicorn", "h11", "h2", "httptools", "websockets"]
 
 [tool.poetry.urls]
 "Issue Tracker" = "https://github.com/emmett-framework/emmett/issues"
